Finished reworking the state threading.
All eval functions now go from state to state * parsetree tuple.
We are also performing a simplification step on the final parsetree where
when a leaf node is enclosed in a subtree node with only one child, and
if the subtree and leaf node agree on all other fields, the subtree node
is substituted with the leave node.

E.e. (simplified)
{
    name = "number";
    pos = 0;
    node = Tree [
        {
            name = "number";
            pos = 0;
            node = Lexeme "1";
        }
    ];
}
gets simplified to
{
    name = "number";
    pos = 0;
    node = Lexeme "1";
}